Inline Frame into vm.rs
